[Fix] 그래프 데이터 매번 갱신하기 수정 완료
